# docs/swagger.yaml
definitions:
  commons.SwaggerApiResponseError:
    properties:
      data: {}
      message:
        type: string
      success:
        default: false
        type: boolean
    type: object
  commons.SwaggerApiResponseSuccessWithData:
    properties:
      data: {}
      message:
        type: string
      success:
        default: true
        type: boolean
    type: object
  commons.SwaggerApiResponseSuccessWithoutData:
    properties:
      data: {}
      message:
        type: string
      success:
        default: true
        type: boolean
    type: object
  models.BranchItemRequest:
    properties:
      added_stock:
        type: integer
      branch_id:
        type: integer
      item_id:
        type: integer
    type: object
  models.BranchItemUpdateRequest:
    properties:
      added_stock:
        type: integer
    type: object
  models.BranchRequest:
    properties:
      address:
        type: string
      name:
        type: string
    type: object
...
  models.ItemRequest:
    properties:
      item_type_id:
        type: integer
      name:
        type: string
      price:
        type: integer
      profit:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple store management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Simple Store Management API
  version: "1.0"
paths:
  /api/branch:
    get:
      consumes:
      - application/json
      description: get all branch
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get All Branch
      tags:
      - branch
    post:
      consumes:
      - application/json
      description: create a new branch
      parameters:
      - description: Branch Request
        in: body
        name: branchRequest
        required: true
        schema:
          $ref: '#/definitions/models.BranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Create Branch
      tags:
      - branch
  /api/branch/{id}:
    delete:
      consumes:
      - application/json
      description: delete a branch
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Delete Branch
      tags:
      - branch
    get:
      consumes:
      - application/json
      description: get branch by id
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Branch
      tags:
      - branch
    put:
      consumes:
      - application/json
      description: update a branch
      parameters:
      - description: Branch Request
        in: body
        name: branchRequest
        required: true
        schema:
          $ref: '#/definitions/models.BranchRequest'
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Update Branch
      tags:
      - branch
...
  /api/branch/items/{id}:
    get:
      consumes:
      - application/json
      description: get branch with Iitems
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Branch With Items
      tags:
      - branch
...
  /api/branchItem:
    get:
      consumes:
      - application/json
      description: get all branch items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get All Branch Items
      tags:
      - branchItem
    post:
      consumes:
      - application/json
      description: create a new branch item
      parameters:
      - description: Branch Item Request
        in: body
        name: branchItemRequest
        required: true
        schema:
          $ref: '#/definitions/models.BranchItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Create Branch Item
      tags:
      - branchItem
  /api/branchItem/{id}:
    delete:
      consumes:
      - application/json
      description: delete a branch item
      parameters:
      - description: Branch Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Delete Branch Item
      tags:
      - branchItem
    get:
      consumes:
      - application/json
      description: get branch item by id
      parameters:
      - description: Branch Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Branch Item
      tags:
      - branchItem
    put:
      consumes:
      - application/json
      description: update a branch item
      parameters:
      - description: Branch Item Update Request
        in: body
        name: branchItemUpdateRequest
        required: true
        schema:
          $ref: '#/definitions/models.BranchItemUpdateRequest'
      - description: Branch Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Update Branch Item
      tags:
      - branchItem
...
  /api/item:
    get:
      consumes:
      - application/json
      description: get all items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get All Items
      tags:
      - item
    post:
      consumes:
      - application/json
      description: create a new item
      parameters:
      - description: Item Request
        in: body
        name: itemRequest
        required: true
        schema:
          $ref: '#/definitions/models.ItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Create Item
      tags:
      - item
  /api/item/{id}:
    delete:
      consumes:
      - application/json
      description: delete item by id
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Delete Item
      tags:
      - item
    get:
      consumes:
      - application/json
      description: get item by id
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Item
      tags:
      - item
    put:
      consumes:
      - application/json
      description: update item by id
      parameters:
      - description: Item Request
        in: body
        name: itemRequest
        required: true
        schema:
          $ref: '#/definitions/models.ItemRequest'
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Update Item
      tags:
      - item
...
  /api/itemType/items/{id}:
    get:
      consumes:
      - application/json
      description: get items of item type
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Items Of Item Type
      tags:
      - itemType
...
  /api/salesData/branch/{branchId}:
    get:
      consumes:
      - application/json
      description: get sales data by branch
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Sales Data Branch
      tags:
      - salesData
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"